ENV/requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
llvmlite==0.39.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.5
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
widgetsnbextension==4.0.7
